Move functions for reading and writing trk files into trk_io
