+++ deepcell_tracking/__init__.py
@@ -31,6 +31,7 @@ from __future__ import print_function
 from deepcell_tracking.tracking import CellTracker
 from deepcell_tracking import isbi_utils
 from deepcell_tracking import utils
+from deepcell_tracking import trk_io
 
 del absolute_import
 del division
